+++ tests_integration/serviceClientTokenProvider.test.js
@@ -8,11 +8,9 @@ beforeEach(() => { sandbox = sinon.createSandbox(); });
 afterEach(() => sandbox.restore());
 
 describe('serviceClientTokenProvider.js', () => {
-  describe('tokenProvider as a function itself', () => {
-    it('Validate cache tokens work', async () => {
-      const accessKey = 'ACCESS_KEY';
-      const authressClient = new AuthressClient({ authressApiUrl: 'API_URL' }, accessKey);
-      await authressClient.serviceClients.createClient({ clientId: 'Something' });
-    });
+  it('Create a service client', async () => {
+    const accessKey = 'ACCESS_KEY';
+    const authressClient = new AuthressClient({ authressApiUrl: 'API_URL' }, accessKey);
+    await authressClient.serviceClients.createClient({ name: 'Javascript SDK Integration Test Client', clientId: 'Something' });
   });
 });